backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning & Forecasting
scikit-learn==1.3.2
statsmodels==0.14.0
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Optimization Algorithms
pulp==2.7.0
ortools==9.8.3296

# Data Processing
openpyxl==3.1.2
python-dateutil==2.8.2

# Configuration
pyyaml==6.0.1

# Logging & Utilities
loguru==0.7.2

# Parallel Processing
joblib==1.3.2

# Geospatial / Routing Support
geopy==2.4.1

# Security & Environment
python-dotenv==1.0.1

# API Client Calls (optional but common)
requests==2.31.0
